Fixed machine removal and helper bug.

+++ bsHorizonAdSync.ps1
@@ -18,61 +18,62 @@ $Global:HzPass   = ConvertTo-SecureString $settingsFile[3].Trim() -AsPlainText -
 
 <#-----------------------------------------------------------------------------------------------------------------------------+
 
- - This script syncs Active Directory computer groups with VMWare Horizon MANUAL desktop pools.
+    - This script syncs Active Directory computer groups with VMWare Horizon MANUAL desktop pools.
 
- - This script DOES NOT remove pools or entitlements if their source AD group is removed.
- 
- - Finds AD groups with the prefix setting, creates a matching Horizon pool and global entitlement, 
- - Gets computers from the AD group, including any nested groups, and adds the ones registered in Horizon into the pool.
- - Gives permission on the global entitlement if a matching AD user access group exists.
- - Updates the display name for the pool and global entitlement as long as the pool name remains the same.
- - Although coded to generate a display name from the group name, it could be loaded from a separate group attribute.
+    - This script DOES NOT remove pools or entitlements if their source AD group is removed.
 
- - By Richard Perry, Last updated 2020-09-03.
+    - Finds AD groups with the prefix setting, creates a matching Horizon pool and global entitlement,
+    - Gets computers from the AD group, including any nested groups, and adds the ones registered in Horizon into the pool.
+    - Gives permission on the global entitlement if a matching AD user access group exists.
+    - Updates the display name for the pool and global entitlement as long as the pool name remains the same.
+    - Although coded to generate a display name from the group name, it could be loaded from a separate group attribute.
 
--------------------------------------------------------------------------------------------------------------------------------+
-
- ONE TIME ENVIRONMENT SETUP:
-
- [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
- Install-Module VMware.PowerCLI
-
- Set-PowerCLIConfiguration -ParticipateInCEIP $false        -Scope AllUsers -Confirm:$false
- Set-PowerCLIConfiguration -InvalidCertificateAction Ignore -Scope AllUsers -Confirm:$false
-
- Download and extract whole project zip from: https://github.com/vmware/PowerCLI-Example-Scripts
- Copy 'VMware.Hv.Helper' folder from 'PowerCLI-Example-Scripts-master\Modules' into 'C:\Program Files\WindowsPowerShell\Modules'
-
- Close any PowerShell windows or ISE and re-open so changes and modules are loaded.
+    - By Richard Perry, Last updated 2020-09-17.
 
 -------------------------------------------------------------------------------------------------------------------------------+
 
- CODE AND VMWARE.HV.HELPER 1.3.1 NOTES:
+    ONE TIME ENVIRONMENT SETUP:
 
- Get-ADGroupMember is limited by DC setting 'MaxGroupOrMemberEntries' which by default is 5000 members returned.
+    [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
+    Install-Module VMware.PowerCLI
 
- Use the Horizon 'xxxSummary' commands where possible as they are quicker and easier for getting information.
+    Set-PowerCLIConfiguration -ParticipateInCEIP $false        -Scope AllUsers -Confirm:$false
+    Set-PowerCLIConfiguration -InvalidCertificateAction Ignore -Scope AllUsers -Confirm:$false
 
- The 'New-HVPool' param -VM is the machines to be added, at least one must be provided as empty new pools are not allowed.
- If a machine specified is not registered or used by another pool, it gives a very unhelpful error:
- Exception calling "Desktop_Create" with "2" argument(s): "There is an error in the XML document."
+    Download and extract whole project zip from: https://github.com/vmware/PowerCLI-Example-Scripts
+    Copy 'VMware.Hv.Helper' folder from 'PowerCLI-Example-Scripts-master\Modules' into 'C:\Program Files\WindowsPowerShell\Modules'
 
- The 'New-HVPool' param -GlobalEntitlement does not seem to work, instead use a delay and 'Set-HVPool' after creation.
+    Close any PowerShell windows or ISE and re-open so changes and modules are loaded.
 
- The 'New-HVPool' param -AutomaticLogoffPolicy has mistakenly been restricted to CLONE type pools, so errors on MANUAL.
- Also 'Set-HVPool' does not have AutomaticLogoffPolicy, or support setting its Key, so we use our own helper function.
+-------------------------------------------------------------------------------------------------------------------------------+
 
- The 'Set-HVPool' param -Key is cAsE SeNsItiVe, for example 'base.displayname' instead of 'base.displayName' gives:
- Exception calling "Desktop_Update" with "3" argument(s): "ExceptionType : VMware.Hv.InvalidArgument
+    CODE AND VMWARE.HV.HELPER 1.3.1 NOTES:
 
- The 'Set-HVPool' param -PoolName is optional and will silently return nothing if there was no pool specified.
+    Get-ADGroupMember is limited by DC setting 'MaxGroupOrMemberEntries' which by default is 5000 members returned.
 
- The 'Get-HVPoolSummary' has 'DesktopSummaryData.NumMachine' but it does not update after adding or removing machines,
- only after re-connecting. Instead array wrap 'Get-HVMachineSummary' and count the number of entries.
+    The 'Get-HVMachineSummary' was found buggy, where after time it starts returning partial results like a limit had been set,
+    and required a connection service restart to fix. Instead 'Get-HVQueryResult' is used which seems unaffected by the issue.
 
- The 'Add-HVDesktop' adds machines to an existing pool, but HV.Helper has no 'Remove' or 'Move' function.
- There is a tempting 'Remove-HVMachine' but this unregisters physical desktops from the inventory database.
- Good news is the low-level API does contain a remove from pool option, so we use our own helper function.
+    The 'New-HVPool' param -VM is the machines to be added, at least one must be provided as empty new pools are not allowed.
+    If a machine specified is not registered or used by another pool, it gives a very unhelpful error:
+    Exception calling "Desktop_Create" with "2" argument(s): "There is an error in the XML document."
+
+    The 'New-HVPool' param -GlobalEntitlement does not seem to work, instead use a delay and 'Set-HVPool' after creation.
+
+    The 'New-HVPool' param -AutomaticLogoffPolicy has mistakenly been restricted to CLONE type pools, so errors on MANUAL.
+    Also 'Set-HVPool' does not have AutomaticLogoffPolicy, or support setting its Key, so we use our own helper function.
+
+    The 'Set-HVPool' param -Key is cAsE SeNsItiVe, for example 'base.displayname' instead of 'base.displayName' gives:
+    Exception calling "Desktop_Update" with "3" argument(s): "ExceptionType : VMware.Hv.InvalidArgument
+
+    The 'Set-HVPool' param -PoolName is optional and will silently return nothing if there was no pool specified.
+
+    The 'Get-HVPoolSummary' has 'DesktopSummaryData.NumMachine' but it does not update after adding or removing machines,
+    only after re-connecting. Instead array wrap 'Get-HVMachineSummary' and count the number of entries.
+
+    The 'Add-HVDesktop' adds machines to an existing pool, but HV.Helper has no 'Remove' or 'Move' function.
+    There is a tempting 'Remove-HVMachine' but this unregisters physical desktops from the inventory database.
+    Good news is the low-level API does contain a remove from pool option, so we use our own helper function.
 
 ------------------------------------------------------------------------------------------------------------------------------#>
 
@@ -105,13 +106,6 @@ function Get-HVAdGroupMachines()
 
             # Directing foreach output into a variable builds an array without the overhead of adding.
             $machineDns
-        }
-
-        # Skip AD groups that are empty or only contain duplicates.
-        if (!$adMachines)
-        {
-            "No valid machines in AD group, $($adGroup.Name)" | LogMessage -Color Red
-            continue
         }
 
         # Display name is what people see, can change, and could be a custom attribute.
@@ -198,37 +192,64 @@ function Sync-HVPool()
         $DisplayName,
 
         [Parameter(Mandatory = $true)]
+        [AllowNull()]
         [string[]]
         $Machines
     )
 
-    "Syncing pool, $PoolName" | LogMessage -Color Yellow
+    # If machines to sync is null, change it to an empty array.
+    if (!$Machines) { $Machines = @() }
 
-    # First we remove machines from different pools and filter machines not registered.
+    "Syncing pool, $PoolName, $($Machines.Count)" | LogMessage -Color Yellow
+
+    # Get a list of current pool machines and remove any that should no longer be in it.
+    $poolMachines = $Global:HVMachinesInPool.GetEnumerator() | ForEach-Object { if ($_.Value -eq $PoolName) { $_.Name } }
+    foreach ($machineName in $poolMachines)
+    {
+        # Uses contains operator instead of method as it needs to be case insensitive.
+        if ($Machines -notcontains $machineName)
+        {
+            "Machine no longer in a pool, $machineName" | LogMessage
+            Remove-HVDesktopFromPool -Machines $machineName
+
+            $Global:HVMachinesInPool.Remove($machineName)
+            $Global:HVMachinesNonPool.Add($machineName, 'POOL_SYNC')
+        }
+    }
+
+    # Remove machines in a different pool so they can be added to the new pool.
+    foreach ($machineName in $Machines)
+    {
+        if ($Global:HVMachinesInPool.ContainsKey($machineName) -and $Global:HVMachinesInPool[$machineName] -ne $PoolName)
+        {
+            "Machine moving from another pool, $machineName" | LogMessage
+            Remove-HVDesktopFromPool -Machines $machineName
+
+            $Global:HVMachinesInPool.Remove($machineName)
+            $Global:HVMachinesNonPool.Add($machineName, 'POOL_SYNC')
+        }
+    }
+
+    # Make a list of machines to add that are registered in the inventory.
     $machineNamesToAdd = foreach ($machineName in $Machines)
     {
-        if ($Global:HVMachinesNonPool.ContainsKey($machineName))
+        if (!$Global:HVMachinesInPool.ContainsKey($machineName) -and !$Global:HVMachinesNonPool.ContainsKey($machineName))
         {
-            "Machine available, $machineName" | LogMessage
-            $machineName
+            "Machine not registered, $machineName" | LogMessage
             continue
         }
 
-        if ($Global:HVMachinesInPool.ContainsKey($machineName))
+        if ($Global:HVMachinesInPool.ContainsKey($machineName) -and $Global:HVMachinesInPool[$machineName] -eq $PoolName)
         {
-            if ($Global:HVMachinesInPool[$machineName] -eq $PoolName)
-            {
-                "Machine already in pool, $machineName" | LogMessage
-                continue
-            }
-
-            "Machine in different pool, $machineName, $($Global:HVMachinesInPool[$machineName])" | LogMessage
-            Remove-HVDesktopFromPool -Machines $machineName
-            $machineName
+            "Machine in correct pool, $machineName" | LogMessage
             continue
         }
 
-        "Machine not registered, $machineName" | LogMessage
+        "Machine to be added, $machineName" | LogMessage
+        $machineName
+
+        $Global:HVMachinesNonPool.Remove($machineName)
+        $Global:HVMachinesInPool.Add($machineName, $PoolName)
     }
 
     # Check if the pool needs to be created or updated.
@@ -265,38 +286,36 @@ function Sync-HVPool()
 
         return
     }
+
+    # Update the pool display name if needed.
+    $poolDisplayName = $poolSummary.DesktopSummaryData.DisplayName
+    if ($poolDisplayName -cne $DisplayName)
+    {
+        "Updating pool display name, $PoolName, $poolDisplayName, $DisplayName" | LogMessage -Color Yellow
+        Set-HVPool -PoolName $PoolName -Key 'base.displayName' -Value $DisplayName
+    }
     else
     {
-        # Update the pool display name if needed.
-        $poolDisplayName = $poolSummary.DesktopSummaryData.DisplayName
-        if ($poolDisplayName -cne $DisplayName)
-        {
-            "Updating pool display name, $PoolName, $poolDisplayName, $DisplayName" | LogMessage -Color Yellow
-            Set-HVPool -PoolName $PoolName -Key 'base.displayName' -Value $DisplayName
-        }
-        else
-        {
-            "Pool display name matches, $DisplayName" | LogMessage -Color Green
-        }
+        "Pool display name matches, $DisplayName" | LogMessage -Color Green
+    }
 
-        # Check if pool is in a global entitlement.
-        if ($poolSummary.DesktopSummaryData.GlobalEntitlement)
+    # Check if pool is in a global entitlement.
+    if ($poolSummary.DesktopSummaryData.GlobalEntitlement)
+    {
+        # Pool summary only has the ID of its global entitlement, so we need to look up the current display name.
+        $geIdId    = $poolSummary.DesktopSummaryData.GlobalEntitlement.Id
+        $geSummary = Get-HVQueryResult -EntityType GlobalEntitlementSummaryView | Where-Object { $_.Id.Id -eq $geIdId }
+        if ($geSummary)
         {
-            # Pool summary only has the ID of its global entitlement, so we need to look up the current display name.
-            $geIdId    = $poolSummary.DesktopSummaryData.GlobalEntitlement.Id
-            $geSummary = Get-HVQueryResult -EntityType GlobalEntitlementSummaryView | Where-Object { $_.Id.Id -eq $geIdId }
-            if ($geSummary)
+            $geDisplayName = $geSummary.Base.DisplayName
+            if ($geDisplayName -cne $DisplayName)
             {
-                $geDisplayName = $geSummary.Base.DisplayName
-                if ($geDisplayName -cne $DisplayName)
-                {
-                    "Updating global entitlement display name, $geDisplayName, $DisplayName" | LogMessage -Color Yellow
-                    Set-HVGlobalEntitlement -displayName $geDisplayName -Key 'base.displayName' -Value $DisplayName
-                }
-                else
-                {
-                    "Global entitlement display name matches, $DisplayName" | LogMessage -Color Green
-                }
+                "Updating global entitlement display name, $geDisplayName, $DisplayName" | LogMessage -Color Yellow
+                Set-HVGlobalEntitlement -displayName $geDisplayName -Key 'base.displayName' -Value $DisplayName
+            }
+            else
+            {
+                "Global entitlement display name matches, $DisplayName" | LogMessage -Color Green
             }
         }
     }
@@ -304,7 +323,7 @@ function Sync-HVPool()
     # Check there are machines to add.
     if (!$machineNamesToAdd)
     {
-        "No changes to sync, $PoolName" | LogMessage -Color Yellow
+        "No machines to sync, $PoolName" | LogMessage -Color Yellow
         return
     }
 
@@ -318,7 +337,7 @@ function Set-HVGlobalMachineList()
     # Put machines that are in a pool or available into a global hashtable for fast lookup.
 
     $Global:HVMachinesInPool = @{}
-    $machineSummary = Get-HVMachineSummary
+    $machineSummary = Get-HVQueryResult -EntityType MachineNamesView
     foreach ($machine in $machineSummary)
     {
         $Global:HVMachinesInPool.Add($machine.Base.Name, $machine.Base.DesktopName)
@@ -365,7 +384,7 @@ function Set-HVPoolLogOff()
     # This prevents running Set-HVPool twice for 'AFTER' and -Spec param only supports external files.
     # So instead we go down a layer and build our own multi-value update entry.
 
-    $poolUpdates = @()        
+    $poolUpdates = @()
 
     $mapEntry        = New-Object VMware.Hv.MapEntry
     $mapEntry.Key    = 'desktopSettings.logoffSettings.automaticLogoffPolicy'
@@ -379,7 +398,7 @@ function Set-HVPoolLogOff()
         $mapEntry.Value  = $AutomaticLogoffMinutes
         $poolUpdates    += $mapEntry
     }
-    
+
     "Updating pool logoff, $PoolName, $AutomaticLogoffPolicy, $AutomaticLogoffMinutes" | LogMessage -Color Magenta
 
     try
@@ -405,7 +424,7 @@ function Remove-HVDesktopFromPool()
 
     foreach ($machineName in $Machines)
     {
-        $machineSummary = Get-HVMachineSummary -MachineName $machineName
+        $machineSummary = Get-HVQueryResult -EntityType MachineNamesView -Filter (Get-HVQueryFilter base.name -eq $machineName)
 
         if (!$machineSummary)
         {
@@ -442,7 +461,10 @@ function Get-HVPoolMachineCount()
         $PoolName
     )
 
-    @(Get-HVMachineSummary -PoolName $PoolName -SuppressInfo $true).Count
+    $poolSummary    = Get-HVQueryResult -EntityType DesktopSummaryView -Filter (Get-HVQueryFilter desktopSummaryData.name -eq $PoolName)
+    $machineSummary = Get-HVQueryResult -EntityType MachineNamesView   -Filter (Get-HVQueryFilter base.desktop -eq $poolSummary.Id)
+
+    @($machineSummary).Count
 }
 
 function LogDelete()
@@ -504,9 +526,13 @@ if (!(Test-Path variable:Global:DefaultHVServers) -or !$Global:DefaultHVServers)
     }
 }
 
-# Get available Horizon machines and start the AD search.
+# Get available Horizon machines.
 Set-HVGlobalMachineList
+"Total pool and non-pool machines, $($Global:HVMachinesInPool.Count), $($Global:HVMachinesNonPool.Count)" | LogMessage -Color Magenta
+
+# Start the AD search and pool sync.
 Get-HVAdGroupMachines
+"Total pool and non-pool machines, $($Global:HVMachinesInPool.Count), $($Global:HVMachinesNonPool.Count)" | LogMessage -Color Magenta
 
 # Actively disconnect from the Horizon server.
 Disconnect-HVServer -Confirm:$false
